refactor(frontend): extract components, centralize colors in theme.ts, remove all inline styles
